+++ FanControl/Models/ErrorAlert.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct ErrorAlert: Identifiable {
+    let id = UUID()
+    let message: String
+}
