+++ FanControl/Models/UpdateStatusAlert.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct UpdateStatusAlert: Identifiable {
+    let id = UUID()
+    let title: String
+    let message: String
+}
