+++ FanControl/Models/UpdateStatusAlertPresenter.swift
@@ -1,0 +1,3 @@
+enum UpdateStatusAlertPresenter {
+    case mainWindow, settings, menuBar
+}
